Personal Website
Initial commit for my personal website.

Plans:

The plan is to have a CV in markdown.
Their is blog section if I ever decide to make a blog.

The only project that I want to show is the Telemedical in the future I
will hopefully have more.

Probably add some animation to this project.